Reorganizing storage samples.
