+++ README.md
@@ -8,16 +8,6 @@ This repository holds the samples used in the python documentation on [cloud.goo
 For more detailed introduction to a product, check the README in the
 corresponding folder.
 
-## Running the samples
-
-Most samples must be run as modules instead of directly, for example:
-
-```
-$ python -m bigquery.samples.async_query [your-project-id] [your-query]
-```
-
-Refer to the README in the corresponding folder for any special instructions.
-
 ## Contributing changes
 
 * See [CONTRIBUTING.md](CONTRIBUTING.md)
