+++ bigquery/README.md
@@ -2,14 +2,6 @@
 
 This section contains samples for [Google BigQuery](https://cloud.google.com/bigquery).
 
-## Running the samples
-
-These samples must be run as modules, for example:
-
-```
-$ python -m bigquery.samples.async_query [your-project-id] [your-query]
-```
-
 ## Other Samples
 
 * [Using BigQuery from Google App Engine](../appengine/bigquery).
