+++ storage/api/test_list_objects.py
@@ -11,7 +11,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-from storage.list_objects import main
+from .list_objects import main
 from tests import CloudBaseTest
