--- storage/storage_transfer/test_nearline_request.py
@@ -1,39 +1,0 @@
-# Copyright 2015, Google, Inc.
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-import unittest
-
-from mock import Mock
-from mock import patch
-
-from nearline_request import main
-
-
-class NearlineRequestTestCase(unittest.TestCase):
-    """A test case for creating a TransferJob to Nearline for old files."""
-
-    def setUp(self):
-        patcher1 = patch(
-            'storage.storage_transfer.nearline_request.create_client')
-        self.mock_create_client = patcher1.start()
-        self.addCleanup(patcher1.stop)
-        self.mock_client = Mock(spec=['transferJobs'])
-        self.mock_create_client.create_transfer_client.return_value = \
-            self.mock_client
-
-    def test_create_nearline_request(self):
-        execute = self.mock_client.transferJobs.return_value.create.return_value \
-            .execute
-        execute.return_value = ""
-        main()
-        execute.assert_called_with()
